Move WebSocket to websocket package
